test(qa): extract test utilities
