+++ qa/integration-tests/src/test/java/io/camunda/zeebe/it/clustering/dynamic/PartitionJoinTest.java
@@ -7,17 +7,17 @@
  */
 package io.camunda.zeebe.it.clustering.dynamic;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import static io.camunda.zeebe.it.clustering.dynamic.Utils.assertChangeIsApplied;
+import static io.camunda.zeebe.it.clustering.dynamic.Utils.assertChangeIsCompleted;
+import static io.camunda.zeebe.it.clustering.dynamic.Utils.assertChangeIsPlanned;
 
 import io.camunda.zeebe.management.cluster.PostOperationResponse;
-import io.camunda.zeebe.management.cluster.TopologyChange.StatusEnum;
 import io.camunda.zeebe.qa.util.actuator.ClusterActuator;
 import io.camunda.zeebe.qa.util.cluster.TestCluster;
-import java.time.OffsetDateTime;
 import java.util.stream.Stream;
+import org.awaitility.Awaitility;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
-import org.testcontainers.shaded.org.awaitility.Awaitility;
 
 final class PartitionJoinTest {
   @ParameterizedTest
@@ -33,34 +33,6 @@ final class PartitionJoinTest {
           .untilAsserted(() -> assertChangeIsCompleted(cluster, response));
       assertChangeIsApplied(cluster, response);
     }
-  }
-
-  private void assertChangeIsPlanned(final PostOperationResponse response) {
-    assertThat(response.getPlannedChanges()).isNotEmpty();
-    assertThat(response.getExpectedTopology())
-        .usingRecursiveComparison()
-        .ignoringFieldsOfTypes(OffsetDateTime.class)
-        .isNotEqualTo(response.getCurrentTopology());
-  }
-
-  private void assertChangeIsApplied(
-      final TestCluster cluster, final PostOperationResponse response) {
-    final var actuator = ClusterActuator.of(cluster.availableGateway());
-    final var expectedTopology = response.getExpectedTopology();
-    final var currentTopology = actuator.getTopology().getBrokers();
-    assertThat(currentTopology)
-        .usingRecursiveComparison()
-        .ignoringFields("lastUpdatedAt")
-        .isEqualTo(expectedTopology);
-  }
-
-  private void assertChangeIsCompleted(
-      final TestCluster cluster, final PostOperationResponse response) {
-    final var actuator = ClusterActuator.of(cluster.availableGateway());
-    final var currentChange = actuator.getTopology().getChange();
-    assertThat(currentChange).isNotNull();
-    assertThat(currentChange.getId()).isEqualTo(response.getChangeId());
-    assertThat(currentChange.getStatus()).isEqualTo(StatusEnum.COMPLETED);
   }
 
   static Stream<Scenario> testScenarios() {
