+++ qa/integration-tests/src/test/java/io/camunda/zeebe/it/clustering/dynamic/PartitionLeaveTest.java
@@ -7,16 +7,17 @@
  */
 package io.camunda.zeebe.it.clustering.dynamic;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import static io.camunda.zeebe.it.clustering.dynamic.Utils.assertChangeIsApplied;
+import static io.camunda.zeebe.it.clustering.dynamic.Utils.assertChangeIsCompleted;
+import static io.camunda.zeebe.it.clustering.dynamic.Utils.assertChangeIsPlanned;
 
 import io.camunda.zeebe.management.cluster.PostOperationResponse;
-import io.camunda.zeebe.management.cluster.TopologyChange.StatusEnum;
 import io.camunda.zeebe.qa.util.actuator.ClusterActuator;
 import io.camunda.zeebe.qa.util.cluster.TestCluster;
 import java.util.stream.Stream;
+import org.awaitility.Awaitility;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
-import org.testcontainers.shaded.org.awaitility.Awaitility;
 
 final class PartitionLeaveTest {
   @ParameterizedTest
@@ -27,30 +28,11 @@ final class PartitionLeaveTest {
       // when
       final var response = runOperation(cluster, scenario.operation());
       // then
+      assertChangeIsPlanned(response);
       Awaitility.await("Requested change is completed in time")
           .untilAsserted(() -> assertChangeIsCompleted(cluster, response));
       assertChangeIsApplied(cluster, response);
     }
-  }
-
-  private void assertChangeIsApplied(
-      final TestCluster cluster, final PostOperationResponse response) {
-    final var actuator = ClusterActuator.of(cluster.availableGateway());
-    final var expectedTopology = response.getExpectedTopology();
-    final var currentTopology = actuator.getTopology().getBrokers();
-    assertThat(currentTopology)
-        .usingRecursiveComparison()
-        .ignoringFields("lastUpdatedAt")
-        .isEqualTo(expectedTopology);
-  }
-
-  private void assertChangeIsCompleted(
-      final TestCluster cluster, final PostOperationResponse response) {
-    final var actuator = ClusterActuator.of(cluster.availableGateway());
-    final var currentChange = actuator.getTopology().getChange();
-    assertThat(currentChange).isNotNull();
-    assertThat(currentChange.getId()).isEqualTo(response.getChangeId());
-    assertThat(currentChange.getStatus()).isEqualTo(StatusEnum.COMPLETED);
   }
 
   static Stream<Scenario> testScenarios() {
